Disable incremental compilation when KSP is used.
This is a workaround to make sure that processors re-run with proper
inputs when there are changes to processors themselves or their inputs.
Currently KSP lives within the same task with the main compilation. When
it is triggered and what inputs it receives are subject to the logic of
incremental compilation, which is different from the requirement of KSP.